CI: combine lint and test stages (MR 2478)
While there are places where it would make some sense to catch issues in
the lint stage, most of the lint failures are things like shellchcek or
ruff which don't actually effect the functionality of the code.

Let's just run pytest at the same time as these to speed up the overall
pipeline.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>