pmb.netboot: Only attempt to install nbd after rootfs existance check (MR 2540)
The install action will initialise the native chroot, so this way we
don't have to initialise the chroot in this code path. It also gives
users feedback quicker if they haven't already initialised the rootfs
since this way they don't have to wait around for nbd to be installed
before being told that they need to generate the rootfs first.

Closes https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2549