ci/markdown: don't run "npm install" as root (MR 2485)
Move the "npm install" command to the section that runs as regular user
instead of root. Instead of creating a /usr/bin/markdownlint-cli
symlink, run the JS file from the path where npm installs it. Try to
find a markdownlint-cli executable and use it instead if it exists, in
case users run this script outside of "pmbootstrap chroot" and have the
linter installed under that name.